update shader uniforms
update shader uniforms

+++ app/src/main/assets/testfrag_line.glsl
@@ -1,5 +1,5 @@
 #ifdef GL_ES
-precision mediump float;
+precision highp float;
 #endif
 
 uniform float u_time;
@@ -41,4 +41,3 @@ void main()
 
 	gl_FragColor = vec4( color, 1.0 );
 }
-
